code fixes

+++ BotsSDK/korebot/src/main/java/kore/botssdk/charts/charts/PieRadarChartBase.java
@@ -178,8 +178,8 @@ public abstract class PieRadarChartBase<T extends ChartData<? extends IDataSet<?
 
     public float getAngleForPoint(float x, float y) {
         MPPointF c = this.getCenterOffsets();
-        double tx = (double)(x - c.x);
-        double ty = (double)(y - c.y);
+        double tx = x - c.x;
+        double ty = y - c.y;
         double length = Math.sqrt(tx * tx + ty * ty);
         double r = Math.acos(ty / length);
         float angle = (float)Math.toDegrees(r);
@@ -203,8 +203,8 @@ public abstract class PieRadarChartBase<T extends ChartData<? extends IDataSet<?
     }
 
     public void getPosition(MPPointF center, float dist, float angle, MPPointF outputPoint) {
-        outputPoint.x = (float)((double)center.x + (double)dist * Math.cos(Math.toRadians((double)angle)));
-        outputPoint.y = (float)((double)center.y + (double)dist * Math.sin(Math.toRadians((double)angle)));
+        outputPoint.x = (float)((double)center.x + (double)dist * Math.cos(Math.toRadians(angle)));
+        outputPoint.y = (float)((double)center.y + (double)dist * Math.sin(Math.toRadians(angle)));
     }
 
     public float distanceToCenter(float x, float y) {
@@ -224,7 +224,7 @@ public abstract class PieRadarChartBase<T extends ChartData<? extends IDataSet<?
             yDist = c.y - y;
         }
 
-        dist = (float)Math.sqrt(Math.pow((double)xDist, 2.0D) + Math.pow((double)yDist, 2.0D));
+        dist = (float)Math.sqrt(Math.pow(xDist, 2.0D) + Math.pow(yDist, 2.0D));
         MPPointF.recycleInstance(c);
         return dist;
     }
@@ -287,7 +287,7 @@ public abstract class PieRadarChartBase<T extends ChartData<? extends IDataSet<?
     public void spin(int durationmillis, float fromangle, float toangle, Easing.EasingFunction easing) {
         this.setRotationAngle(fromangle);
         ObjectAnimator spinAnimator = ObjectAnimator.ofFloat(this, "rotationAngle", fromangle, toangle);
-        spinAnimator.setDuration((long)durationmillis);
+        spinAnimator.setDuration(durationmillis);
         spinAnimator.setInterpolator(easing);
         spinAnimator.addUpdateListener(new ValueAnimator.AnimatorUpdateListener() {
             public void onAnimationUpdate(ValueAnimator animation) {
