+++ BotsSDK/widgetsdk/src/main/java/com/kore/ai/widgetsdk/multilamp/shapes/RectangleCompos.java
@@ -126,7 +126,7 @@ public class RectangleCompos implements Shape {
                 txtPaint.setTextAlign(Paint.Align.RIGHT);
                 float textly = point.y + (target.getView().getHeight() / 2) + (6 * density);
                 if (lines.length > 0) {
-                    textly = textly - ((22 * ((int) lines.length / 2) * density));
+                    textly = textly - ((22 * (lines.length / 2) * density));
                 }
                 for (int line = 0; line < lines.length; line++) {
                     Log.d("texty", String.valueOf(textly));
@@ -153,7 +153,7 @@ public class RectangleCompos implements Shape {
 
                 float textry = point.y + (target.getView().getHeight() ) + (6 * density);
                 if (lines.length > 0) {
-                    textry = textry - ((22 * ((int) lines.length / 2) * density));
+                    textry = textry - ((22 * (lines.length / 2) * density));
                 }
                 for (int line = 0; line < lines.length; line++) {
                     Log.d("texty", String.valueOf(textry));
